+++ imod/mf6/model.py
@@ -23,9 +23,7 @@ from imod.mf6.interfaces.imodel import IModel
 from imod.mf6.package import Package
 from imod.mf6.statusinfo import NestedStatusInfo, StatusInfo, StatusInfoBase
 from imod.mf6.utilities.mask import _mask_all_packages
-from imod.mf6.utilities.regrid import (
-    _regrid_like,
-)
+from imod.mf6.utilities.regrid import RegridderWeightsCache, _regrid_like
 from imod.mf6.validation import pkg_errors_to_status_info
 from imod.mf6.write_context import WriteContext
 from imod.schemata import ValidationError
@@ -468,7 +466,10 @@ class Modflow6Model(collections.UserDict, IModel, abc.ABC):
         return clipped
 
     def regrid_like(
-        self, target_grid: GridDataArray, validate: bool = True
+        self,
+        target_grid: GridDataArray,
+        validate: bool = True,
+        regrid_context: Optional[RegridderWeightsCache] = None,
     ) -> "Modflow6Model":
         """
         Creates a model by regridding the packages of this model to another discretization.
@@ -482,13 +483,16 @@ class Modflow6Model(collections.UserDict, IModel, abc.ABC):
             a grid defined over the same discretization as the one we want to regrid the package to
         validate: bool
             set to true to validate the regridded packages
+        regrid_context: Optional RegridderWeightsCache
+            stores regridder weights for different regridders. Can be used to speed up regridding,
+            if the same regridders are used several times for regridding different arrays.
 
         Returns
         -------
         a model with similar packages to the input model, and with all the data-arrays regridded to another discretization,
         similar to the one used in input argument "target_grid"
         """
-        return _regrid_like(self, target_grid, validate)
+        return _regrid_like(self, target_grid, validate, regrid_context)
 
     def mask_all_packages(
         self,