+++ imod/mf6/package.py
@@ -25,6 +25,7 @@ from imod.mf6.pkgbase import (
 from imod.mf6.utilities.mask import _mask
 from imod.mf6.utilities.regrid import (
     RegridderType,
+    RegridderWeightsCache,
     _regrid_like,
 )
 from imod.mf6.utilities.schemata import filter_schemata_dict
@@ -554,6 +555,7 @@ class Package(PackageBase, IPackage, abc.ABC):
     def regrid_like(
         self,
         target_grid: GridDataArray,
+        regrid_context: RegridderWeightsCache,
         regridder_types: Optional[dict[str, Tuple[RegridderType, str]]] = None,
     ) -> "Package":
         """
@@ -580,6 +582,9 @@ class Package(PackageBase, IPackage, abc.ABC):
         regridder_types: dict(str->(regridder type,str))
            dictionary mapping arraynames (str) to a tuple of regrid type (a specialization class of BaseRegridder) and function name (str)
             this dictionary can be used to override the default mapping method.
+        regrid_context: Optional RegridderWeightsCache
+            stores regridder weights for different regridders. Can be used to speed up regridding,
+            if the same regridders are used several times for regridding different arrays.
 
         Returns
         -------
@@ -587,7 +592,7 @@ class Package(PackageBase, IPackage, abc.ABC):
         similar to the one used in input argument "target_grid"
         """
         try:
-            result = _regrid_like(self, target_grid, regridder_types)
+            result = _regrid_like(self, target_grid, regrid_context, regridder_types)
         except ValueError as e:
             raise e
         except Exception: