+++ imod/tests/test_mf6/test_mf6_regrid_model.py
@@ -113,8 +113,9 @@ def test_model_regridding_can_skip_validation(
     """
 
     # create a sto package with a negative storage coefficient. This would trigger a validation error if it were turned on.
-    storage_coefficient = grid_data_structured(np.float64, -20.0, 0.25)
-    specific_yield = grid_data_structured(np.float64, -30.0, 0.25)
+    storage_coefficient = grid_data_structured(np.float64, -20.0, 2.0)
+    specific_yield = grid_data_structured(np.float64, -30.0, 2.0)
+
     sto_package = imod.mf6.StorageCoefficient(
         storage_coefficient,
         specific_yield,
@@ -162,8 +163,8 @@ def test_model_regridding_can_validate(
     """
 
     # Create a storage package with a negative storage coefficient. This would trigger a validation error if it were turned on.
-    storage_coefficient = grid_data_structured(np.float64, -20, 0.25)
-    specific_yield = grid_data_structured(np.float64, -30, 0.25)
+    storage_coefficient = grid_data_structured(np.float64, -20, 2.0)
+    specific_yield = grid_data_structured(np.float64, -30, 2.0)
     sto_package = imod.mf6.StorageCoefficient(
         storage_coefficient,
         specific_yield,