+++ imod/tests/test_mf6/test_mf6_unsupported_grid_operations.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
+from imod.mf6.utilities.regrid import RegridderWeightsCache
 from imod.typing.grid import zeros_like
 
 
@@ -55,10 +56,11 @@ def test_mf6_model_regrid_with_lakes(rectangle_with_lakes, tmp_path):
 def test_mf6_package_regrid_with_lakes(rectangle_with_lakes, tmp_path):
     simulation = rectangle_with_lakes
     package = simulation["GWF_1"]["lake"]
+    old_grid = simulation["GWF_1"].domain
     new_grid = finer_grid(simulation["GWF_1"].domain)
-
+    regrid_context = RegridderWeightsCache(old_grid.sel(layer=1), new_grid.sel(layer=1))
     with pytest.raises(ValueError, match="package(.+)not be regridded"):
-        _ = package.regrid_like(new_grid)
+        _ = package.regrid_like(new_grid, regrid_context)
 
 
 @pytest.mark.usefixtures("rectangle_with_lakes")